56_DAY_PRACTICE/day10_cloud_storage/gemini-function/requirements.txt
functions-framework==3.*
google-genai==1.64.0
